# backend/docs/swagger.yaml
basePath: /api
definitions:
  internal_billing.AdminInstallmentDTO:
    properties:
      amount:
        type: number
      description:
        type: string
      due_date:
        type: string
      id:
        type: string
      status:
        type: string
      student_email:
        type: string
      student_name:
        type: string
    type: object
  internal_billing.CreateInstallmentReq:
    properties:
      amount:
        type: number
      description:
        type: string
      due_date:
        description: 'Formato esperado: YYYY-MM-DD'
        type: string
      user_id:
        type: string
    type: object
  internal_user.LoginRequest:
    properties:
      email:
        type: string
      password:
        type: string
    type: object
host: localhost:3000
info:
  contact:
    name: Carlos Pérez
    url: https://xnzperez-portfolio.vercel.app/
  description: Motor financiero multi-tenant para universidades con control de roles
    y prevención de IDOR.
  title: EduPay SaaS API
  version: "1.0"
paths:
  /billing/installments:
    get:
      consumes:
      - application/json
      description: Obtiene el listado completo de deudas de la universidad con información
        del estudiante.
      parameters:
      - description: ID de la Universidad (UUID)
        in: header
        name: X-Tenant-ID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Lista de deudas
          schema:
            items:
              $ref: '#/definitions/internal_billing.AdminInstallmentDTO'
            type: array
        "500":
          description: Error interno al consultar la base de datos
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Listar todas las deudas (Cajeros)
      tags:
      - Facturación (Cajeros)
    post:
      consumes:
      - application/json
      description: Permite a un cajero asignar una nueva obligación financiera a un
        estudiante. Validado contra fechas pasadas.
      parameters:
      - description: ID de la Universidad (UUID)
        in: header
        name: X-Tenant-ID
        required: true
        type: string
      - description: Datos de la nueva cuota (Monto, Fecha, Descripción)
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_billing.CreateInstallmentReq'
      produces:
      - application/json
      responses:
        "201":
          description: Cuota generada exitosamente
          schema:
            additionalProperties: true
            type: object
        "400":
          description: JSON o fecha inválida
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Error interno al procesar transacción RLS
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Crear una nueva deuda/cuota
      tags:
      - Facturación (Cajeros)
  /billing/installments/{id}/pay:
    post:
      consumes:
      - application/json
      description: Descuenta el saldo de la billetera del estudiante para pagar una
        cuota. Calcula interés de mora dinámico si está vencida.
      parameters:
      - description: ID de la Universidad (UUID)
        in: header
        name: X-Tenant-ID
        required: true
        type: string
      - description: ID de la Cuota (UUID)
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: La cuota ha sido saldada y el registro actualizado
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Fondos insuficientes o cuota ya pagada
          schema:
            additionalProperties: true
            type: object
      security:
      - BearerAuth: []
      summary: Pagar una cuota o deuda
      tags:
      - Facturación (Estudiantes)
  /users/login:
    post:
      consumes:
      - application/json
      description: Autentica a un usuario validando sus credenciales y devuelve un
        token JWT.
      parameters:
      - description: ID de la Universidad (UUID)
        in: header
        name: X-Tenant-ID
        required: true
        type: string
      - description: Credenciales del usuario (Email y Contraseña)
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/internal_user.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Login exitoso con token
          schema:
            additionalProperties: true
            type: object
        "400":
          description: JSON inválido
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Credenciales incorrectas
          schema:
            additionalProperties: true
            type: object
      summary: Iniciar sesión en el sistema
      tags:
      - Autenticación
securityDefinitions:
  BearerAuth:
    description: Escribe 'Bearer ' seguido de tu token JWT.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"